+++ vta-service/src/config.rs
@@ -570,6 +570,7 @@ impl AppConfig {
                     mediator_url: url,
                     mediator_did: did,
                     mediator_host: None,
+                    setup_acl: false,
                 });
             }
             (Ok(url), Err(_)) => {
@@ -577,6 +578,7 @@ impl AppConfig {
                     mediator_url: String::new(),
                     mediator_did: String::new(),
                     mediator_host: None,
+                    setup_acl: false,
                 });
                 messaging.mediator_url = url;
             }
@@ -585,6 +587,7 @@ impl AppConfig {
                     mediator_url: String::new(),
                     mediator_did: String::new(),
                     mediator_host: None,
+                    setup_acl: false,
                 });
                 messaging.mediator_did = did;
             }