+++ vtc-service/src/config.rs
@@ -925,6 +925,7 @@ impl AppConfig {
                     mediator_url: url,
                     mediator_did: did,
                     mediator_host: None,
+                    setup_acl: false,
                 });
             }
             (Ok(url), Err(_)) => {
@@ -932,6 +933,7 @@ impl AppConfig {
                     mediator_url: String::new(),
                     mediator_did: String::new(),
                     mediator_host: None,
+                    setup_acl: false,
                 });
                 messaging.mediator_url = url;
             }
@@ -940,6 +942,7 @@ impl AppConfig {
                     mediator_url: String::new(),
                     mediator_did: String::new(),
                     mediator_host: None,
+                    setup_acl: false,
                 });
                 messaging.mediator_did = did;
             }